\name{facet_gridx}
\alias{FacetGridx}
\alias{facet_gridx}
\title{Lay out panels in a rectangular/tabular manner, with added options}
\description{
  This is an enhanced version of ggplot2's
  \code{facet_grid}.  It adds options to control whether
  the facet rows and/or columns are associated with the
  axes, and displayed on the left/bottom rather than the
  right/top.
}

