\docType{package}
\name{ggplothacks}
\alias{"package-ggplothacks"}
\alias{ggplothacks}
\title{Add-ons and enhancements for ggplot2}
\description{
  This package contains some miscellaneous add-ons for
  ggplot2, including new geoms, new positioning algorithms,
  and new faceting.
}
\details{
  As suggested by the name, these are \emph{hacks}: they
  may not be well thought-out, and are certainly not
  cleanly implemented.
}

