\name{grobGridCols}
\alias{grobGridCols}
\title{Select one or more columns from a grobGrid}
\usage{
  grobGridCols(grid, cols)
}
\arguments{
  \item{grid}{a grobGrid}

  \item{cols}{a vector of one or more columns}
}
\value{
  a grobGrid containing the specified columns
}
\description{
  Select one or more columns from a grobGrid
}

