\name{grobGridRows}
\alias{grobGridRows}
\title{Select one or more rows from a grobGrid}
\usage{
  grobGridRows(grid, rows)
}
\arguments{
  \item{grid}{a grobGrid}

  \item{rows}{a vector of one or more rows}
}
\value{
  a grobGrid containing the specified row
}
\description{
  Select one or more rows from a grobGrid
}

